+++ driver/CMakeFiles/nytro.dir/cmake_clean.cmake
@@ -2,8 +2,10 @@ file(REMOVE_RECURSE
   "../bin/nytro"
   "../bin/nytro.pdb"
   "CMakeFiles/nytro.dir/link.d"
-  "CMakeFiles/nytro.dir/driver.cpp.o"
-  "CMakeFiles/nytro.dir/driver.cpp.o.d"
+  "CMakeFiles/nytro.dir/src/config_loader.cpp.o"
+  "CMakeFiles/nytro.dir/src/config_loader.cpp.o.d"
+  "CMakeFiles/nytro.dir/src/driver.cpp.o"
+  "CMakeFiles/nytro.dir/src/driver.cpp.o.d"
 )
 
 # Per-language clean rules from dependency scanning.