+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Feat: Add basic lua project configuration.
 - Fix: Fixed a bug where functions would continue executing code after a return statement if the return was nested inside a conditional block.
 - Fix: Updated floats to have a precision of 6, doubles to 15.
+- Fix: Added a help menu with the flags `--help`/`-h`.
 
 ## 0.121
 - Feat: Add support for parenthesis in assembly statements `asm();` instead of only braces `asm {}`.